requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
GitPython==3.1.40
kubernetes==28.1.0
gunicorn==21.2.0
